# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/telepulesek": {
      "get": {
        "operationId": "TelepulesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A települések listája sikeresen lekérdezve.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateTelepulesDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Összes település lekérdezése",
        "tags": [
          "Települések"
        ]
      },
      "post": {
        "operationId": "TelepulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTelepulesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A település sikeresen létrehozva.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTelepulesDto"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen bemenet."
          }
        },
        "summary": "Új település létrehozása",
        "tags": [
          "Települések"
        ]
      }
    },
    "/telepulesek/{id}": {
      "get": {
        "operationId": "TelepulesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "A település azonosítója",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A település sikeresen lekérdezve.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTelepulesDto"
                }
              }
            }
          },
          "404": {
            "description": "A település nem található."
          }
        },
        "summary": "Egy település lekérdezése ID alapján",
        "tags": [
          "Települések"
        ]
      },
      "put": {
        "operationId": "TelepulesController_put",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "A település azonosítója",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTelepulesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A település sikeresen frissítve.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTelepulesDto"
                }
              }
            }
          },
          "404": {
            "description": "A település nem található."
          }
        },
        "summary": "Település teljes frissítése ID alapján",
        "tags": [
          "Települések"
        ]
      },
      "patch": {
        "operationId": "TelepulesController_patch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "A település azonosítója",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTelepulesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A település sikeresen frissítve.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTelepulesDto"
                }
              }
            }
          },
          "404": {
            "description": "A település nem található."
          }
        },
        "summary": "Település részleges frissítése ID alapján",
        "tags": [
          "Települések"
        ]
      },
      "delete": {
        "operationId": "TelepulesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "A település azonosítója",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A település sikeresen törölve."
          },
          "404": {
            "description": "A település nem található."
          }
        },
        "summary": "Település törlése ID alapján",
        "tags": [
          "Települések"
        ]
      }
    },
    "/telepulesek/iranyitoszam/{iranyitoszam}": {
      "get": {
        "operationId": "TelepulesController_findIranyitoszam",
        "parameters": [
          {
            "name": "iranyitoszam",
            "required": true,
            "in": "path",
            "description": "A település irányítószáma",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A település sikeresen lekérdezve.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTelepulesDto"
                }
              }
            }
          },
          "404": {
            "description": "A település nem található."
          }
        },
        "summary": "Település lekérdezése irányítószám alapján",
        "tags": [
          "Települések"
        ]
      }
    },
    "/tesztadatok/szemelyekletrehozasa": {
      "post": {
        "description": "Létrehoz egy megadott számú teszt személyt.",
        "operationId": "TesztAdatokController_szemelyekLetrehozasa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "count": {
                    "type": "number",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Teszt személyek sikeresen létrehozva"
          },
          "400": {
            "description": "Érvénytelen adatok"
          }
        },
        "summary": "Személyek tesztadatok létrehozása",
        "tags": [
          "Tesztadatok"
        ]
      }
    },
...
    "/tesztadatok/tesztadatokletrehozasa": {
      "post": {
        "description": "Létrehoz minden tesztadatot (személyek, települések, hírek, események, feliratkozások).",
        "operationId": "TesztAdatokController_tesztAdatokLetrehozasa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "count": {
                    "type": "number",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tesztadatok sikeresen létrehozva"
          }
        },
        "summary": "Összes tesztadat létrehozása",
        "tags": [
          "Tesztadatok"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS keretrendszerrel készített REST API",
    "description": "REST API személyek, címek, események, feliratkozások, hírek és települések kezelésére",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTelepulesDto": {
        "type": "object",
        "properties": {
          "telepules_varos": {
            "type": "string",
            "description": "A település neve"
          },
          "telepules_iranyitoszam": {
            "type": "string",
            "description": "A település irányítószáma"
          },
          "telepules_megye": {
            "type": "string",
            "description": "A település megyéje (opcionális)",
            "nullable": true
          }
        },
        "required": [
          "telepules_varos",
          "telepules_iranyitoszam",
          "telepules_megye"
        ]
      },
      "UpdateTelepulesDto": {
        "type": "object",
        "properties": {
          "telepules_varos": {
            "type": "string",
            "description": "A település neve (opcionális)"
          },
          "telepules_iranyitoszam": {
            "type": "string",
            "description": "A település irányítószáma (opcionális)"
          },
          "telepules_megye": {
            "type": "string",
            "description": "A település megyéje (opcionális)",
            "nullable": true
          }
        }
      }
    }
  }
}